Added extra employee fee input option for each builder inside builders listing page inside admin panel
