+++ src/admin/validators/create-update-extra-fee.ts
@@ -11,5 +11,5 @@ export class CreateUpdateExtraFeeDTO {
     @Type(() => Number)
     @IsNumber()
     @IsNotEmpty()
-    userId: number;
+    companyId: number;
 }
